refactor: extract liquid/reflection.rs to astra-evolution (1,786 LOC)
Move the LLM-native self-reflection engine that produces EvolutionProposal objects
to the astra-evolution crate where it logically belongs.

Changes:
- Copy reflection.rs to astra-evolution::reflection (1,786 LOC)
- Add astra-services dependency to astra-evolution
- Register module in lib.rs
- Replace original with re-export stub
- Fix clippy lints (map_or → is_none_or)
